feat: basic update to ActiveEffect for Foundry v14
