+++ module/data/_module.mjs
@@ -4,7 +4,7 @@ export * as advancement from "./advancement/_module.mjs";
 export * as chat from "./chat/_module.mjs";
 export * as combatant from "./combatant/_module.mjs";
 export * as common from "./common/_module.mjs";
-export * as effects from "./effects/_module.mjs";
+export * as effects from "./effect/_module.mjs";
 export * as fields from "./fields/_module.mjs";
 export * as item from "./item/_module.mjs";
 export * as roll from "./roll/_module.mjs";
@@ -12,7 +12,7 @@ export * as other from "./other/_module.mjs";
 export * as thread from "./thread/_module.mjs";
 
 export * as Actor from "./actor/_module.mjs";
-export * as ActiveEffect from "./effects/_module.mjs";
+export * as ActiveEffect from "./effect/_module.mjs";
 export * as ChatMessage from "./chat/_module.mjs";
 export * as Combatant from "./combatant/_module.mjs";
 export * as Item from "./item/_module.mjs";
