+++ module/data/_old_effects/eae-condition.mjs
