+++ module/data/_old_effects/eae-duration.mjs
