+++ module/data/effect/_module.mjs
@@ -1,0 +1,5 @@
+import EarthdawnActiveEffectData from "./eae.mjs";
+
+export {
+  EarthdawnActiveEffectData,
+};
