+++ module/hooks/init.mjs
@@ -19,7 +19,7 @@ import * as enrichers from "../helpers/enrichers.mjs";
 import { staticStatusId } from "../helpers/document.mjs";
 
 const { DocumentSheetConfig } = foundry.applications.apps;
-const { ActiveEffectConfig, CombatantConfig } = foundry.applications.sheets;
+const { CombatantConfig } = foundry.applications.sheets;
 const { Actors, Items, Journal, Scenes } = foundry.documents.collections;
 
 /**
@@ -160,7 +160,7 @@ function registerSheetApps() {
     label:       "ED.Documents.actorSheetEdHorror"
   } );
 
-  DocumentSheetConfig.unregisterSheet( ActiveEffect, "core", ActiveEffectConfig );
+  /* DocumentSheetConfig.unregisterSheet( ActiveEffect, "core", ActiveEffectConfig );
   DocumentSheetConfig.registerSheet(
     ActiveEffect,
     "earthdawn4e",
@@ -169,7 +169,7 @@ function registerSheetApps() {
       makeDefault: true,
       label:       "ED.Documents.activeEffectSheetEd"
     }
-  );
+  ); */
 
   DocumentSheetConfig.unregisterSheet( Combatant, "core", CombatantConfig );
   DocumentSheetConfig.registerSheet(
