+++ system.json
@@ -208,8 +208,7 @@
     },
     "documentTypes": {
         "ActiveEffect": {
-            "eae": {},
-            "condition": {}
+            "eae": {}
         },
         "Actor": {
             "character": {
